Extract dns1123_name into an executable (#456)
* Extract dns1123_name into an executable

* Reference util in provision.py

* Run formatter

* Make make_dns1123_name.py executable
